requirements.txt
django==3.2
psycopg2-binary==2.8.5
djangorestframework==3.12.2
django-rest-authtoken==2.1.3 
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
